chore(karma): launch karma in background via gulp task, improve OS compat
There were issues loading the karma binary on windows, due to the path not being available, and the
executable no longer being exported in package.json. So to get around this, the solution is to
simply launch a child node process which launches a server. This works fairly nicely here.

Fixes #4