fix(dirty_checking): reduce GC pressure in DuplicateMap.clear()
Despite my previous attempt at using the Map.prototype.clear() method, it does appear to be available
in the es6-shim, and is clearly defined in the January 20th 2014 ECMAScript 6 harmony draft. Therefore,
there is no reason to trash and re-create a new Map() (unless the shim decides to behave in this way),
as it should reduce GC pressure to use a clear() method which simply removes references to existing
objects.

Closes #8