feat(notifier): add proof of concept for custom notification
The templating engine wants to use a consistent mechanism for databinding throughout its internal implementation. To enable this we need a way to register for changes in certain HTML elements, rather than using dirty checking to observe them. This is a proof of concept for one way that might be plugged in.