feat(observer): enable a general mechanism for plugging in custom field observers
We need a mechanism for handling different types of object/field observation. The first prototype was geared towards handling certain types of bindings to HTML elements. This is the improved implementation which can handle that scenario as well as O.o and any custom change notification system.
Inteternally, explicit observer related modes for the dirty checker are use so that we can noop when no event is raised. Interaction tests included for ObserverSelector and Observer interaction.